Resolution to Issue 8 plus minor upgrades throughout related to boundary conditions and commenting. From now on minor changes not related to a particular issue should not be included.

+++ LatBo/GridObj_ops_immersedboundary.cpp
@@ -100,11 +100,11 @@ void GridObj::ibm_findsupport(unsigned int ib, unsigned int m) {
 
 
 #ifdef CHEAP_NEAREST_NODE_DETECTION
-	inear = floor( iBody[ib].markers[m].position[0]/dx );
-	jnear = floor( iBody[ib].markers[m].position[1]/dy );
+	inear = (unsigned int)floor( iBody[ib].markers[m].position[0]/dx );
+	jnear = (unsigned int)floor( iBody[ib].markers[m].position[1]/dy );
 
 #if (dims == 3)
-	knear = floor( iBody[ib].markers[m].position[2]/dz );
+	knear = (unsigned int)floor( iBody[ib].markers[m].position[2]/dz );
 #endif
